+++ CHANGELOG.rst
@@ -6,12 +6,14 @@ Changelog for Stingray Controller
 
 Added:
 ^^^^^^
+- Offline FW updating
 - Protocol validation checks on import in Stimulation Studio
 
 Fixed:
 ^^^^^^
 - SW auto-updating
 - Issue with firmware versions always reported as being incompatible
+- Prevent connection to Mantarray instruments
 
 
 0.1.0 (2023-04-13)