+++ CHANGELOG.rst
@@ -1,17 +1,19 @@
 Changelog for Stingray Controller
 =================================
 
-0.1.1 (unreleased)
+0.2.0 (unreleased)
 ------------------
 
 Added:
 ^^^^^^
+- Offline FW updating
 - Protocol validation checks on import in Stimulation Studio
 
 Fixed:
 ^^^^^^
 - SW auto-updating
 - Issue with firmware versions always reported as being incompatible
+- Prevent connection to Mantarray instruments
 
 
 0.1.0 (2023-04-13)